+++ sdk/ai/ai-projects/src/api/connections/operations.ts
@@ -5,20 +5,20 @@ import { AIProjectContext as Client } from "../index.js";
 import {
   Connection,
   connectionDeserializer,
+  ConnectionType,
   _PagedConnection,
   _pagedConnectionDeserializer,
-  ConnectionType,
 } from "../../models/models.js";
+import {
+  PagedAsyncIterableIterator,
+  buildPagedAsyncIterator,
+} from "../../static-helpers/pagingHelpers.js";
+import { expandUrlTemplate } from "../../static-helpers/urlTemplate.js";
 import {
   ConnectionsListOptionalParams,
   ConnectionsGetWithCredentialsOptionalParams,
   ConnectionsGetOptionalParams,
 } from "./options.js";
-import { expandUrlTemplate } from "../../static-helpers/urlTemplate.js";
-import {
-  PagedAsyncIterableIterator,
-  buildPagedAsyncIterator,
-} from "../../static-helpers/pagingHelpers.js";
 import {
   StreamableMethod,
   PathUncheckedResponse,