+++ sdk/ai/ai-projects/src/api/evaluationTaxonomies/index.ts
@@ -1,0 +1,11 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { update, create, $delete, list, get } from "./operations.js";
+export {
+  EvaluationTaxonomiesUpdateOptionalParams,
+  EvaluationTaxonomiesCreateOptionalParams,
+  EvaluationTaxonomiesDeleteOptionalParams,
+  EvaluationTaxonomiesListOptionalParams,
+  EvaluationTaxonomiesGetOptionalParams,
+} from "./options.js";